+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## [unreleased]
 
+### Changed
+
+- Expect a tomlkit-specific error instead of `TypeError` from a custom encoder. ([#355](https://github.com/python-poetry/tomlkit/issues/355))
+
 ### Fixed
 
 - Fix the incompatiblity with 3.13 because of the `datetime.replace()` change. ([#333](https://github.com/python-poetry/tomlkit/issues/333))
 - Revert the change of parsing out-of-order tables. ([#347](https://github.com/python-poetry/tomlkit/issues/347))
+
 
 ## [0.12.5] - 2024-05-08